requirements.txt
numpy==1.26.4
scipy
scikit-learn
pandas
networkx
tqdm
pyyaml
dhg==0.9.4
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
